Split install & logging logic
